[FEAT] Components small fix
