+++ src/styles/Header.module.css
@@ -2,7 +2,7 @@
 
 
 .header {
-    width: 100vw;
+    max-width: 100vw;
     height: 10vh;
 }
